+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.3] - 2025-09-25
+- Simplified write/read registers functions to generic Single and Double functions
+- Added Check functions for write/read registers functions
+- Added error code during init if communication error with MCP23S17
+- Replaced inline macros of registers to enum
+- Changed name of attribute value -> ports in struct MCP23S17_t
+- Replaced CalculateModuleId and CalculatePinId functions to inline macros
+
 ## [1.0.2] - 2025-09-23
 - Fixed bug in REG_OLAT_B value